refactoring and retesting
