+++ src/store/Store.js
@@ -1,11 +1,10 @@
-import crypto from 'crypto-js';
 import mitt from 'mitt';
 
 import { parentCall } from '../lib/parentCall';
+import { createToken } from '../lib/random';
 
 const { localStorage, sessionStorage } = window;
 
-const createToken = () => crypto.lib.WordArray.random(128).toString(crypto.enc.Hex);
 export default class Store {
 	constructor(initialState = {}, { localStorageKey = 'store', dontPersist = [] } = {}) {
 		Object.assign(this, mitt());
